Delete the CFFunction, so I can re-add it
gh-1

My stack update failed, but only in prod, which is highly confusing. It
succeeded in the test account. It should be the same across all
accounts.

Here I'm deleting the CFFunction, the resource that wasn't found on
update, so I can add it with a different identifier.

In the CloudFormation stack, the resource name is:

us-east-1CdkturnkeyirectMasterappCFFunction328CA331

but the resource name in CloudFront Functions is:

us-east-1CdkturnkeirectMasterappCFFunction328CA331

The second one has the "y" missing from "Cdkturnkey"! I have no idea
why.