Try to get the error message about new synth
gh-5

My last build failed because it was trying to deploy with no feature
flags (part of the CDK v2 migration) to an env that had been
bootstrapped with CDK 1.

The error message was:

Error: Not downgrading existing bootstrap stack from version '10' to
version '0'. Use --force to force.

In another project, which was at a higher version of CDK 1, the error
message was more helpful:

Error: Not downgrading existing bootstrap stack from version '10' to
version '0'. Use --force to force or set the
'@aws-cdk/core:newStyleStackSynthesis' feature flag in cdk.json to use
the latest bootstrap version.

I want that error message. But even more than that, I want to make sure
that newStyleStackSynthesis is available in the version of CDK 1 that
I'm on when I try to use it.